# Increases magnet pull time to 9 minutes, increases warning time for salvage magnet, decreases wreck cost and other tweaks to the Salvage Magnet.