--- a/Content.Server/Salvage/Magnet/SalvageMagnetDataComponent.cs
+++ b/Content.Server/Salvage/Magnet/SalvageMagnetDataComponent.cs
@@ -29,13 +29,13 @@ public sealed partial class SalvageMagnetDataComponent : Component
     /// How long salvage will be active for before despawning.
     /// </summary>
     [DataField]
-    public TimeSpan ActiveTime = TimeSpan.FromMinutes(6);
+    public TimeSpan ActiveTime = TimeSpan.FromMinutes(9); //DeltaV: was 6 minutes
 
     /// <summary>
     /// Cooldown between offerings after one ends.
     /// </summary>
     [DataField]
-    public TimeSpan OfferCooldown = TimeSpan.FromMinutes(3);
+    public TimeSpan OfferCooldown = TimeSpan.FromMinutes(2); //DeltaV: was 3 min
 
     /// <summary>
     /// Seeds currently offered