--- a/Content.Shared/Salvage/Magnet/SalvageOffering.cs
+++ b/Content.Shared/Salvage/Magnet/SalvageOffering.cs
@@ -7,5 +7,5 @@ public record struct SalvageOffering : ISalvageMagnetOffering
 {
     public SalvageMapPrototype SalvageMap;
 
-    uint ISalvageMagnetOffering.Cost => 500; // DeltaV: was 1000 (from deltanedas), lowered to 500 to make it less of a grind
+    uint ISalvageMagnetOffering.Cost => 500; // DeltaV: Station gets next to no benefit from you pulling wrecks, force you to mine first.
 }