--- a/Content.Shared/Salvage/SharedSalvageSystem.Magnet.cs
+++ b/Content.Shared/Salvage/SharedSalvageSystem.Magnet.cs
@@ -16,9 +16,9 @@ public abstract partial class SharedSalvageSystem
 
     private readonly Dictionary<ISalvageMagnetOffering, float> _offeringWeights = new()
     {
-        { new AsteroidOffering(), 4.5f },
-        { new DebrisOffering(), 3.5f },
-        { new SalvageOffering(), 2.0f },
+        { new AsteroidOffering(), 2.0f }, // DeltaV: was 4.5f
+        { new DebrisOffering(), 3.5f }, // DeltaV: was 3.5f
+        { new SalvageOffering(), 4.5f }, // DeltaV: was 2.0f
     };
 
     private readonly List<ProtoId<DungeonConfigPrototype>> _asteroidConfigs = new()