<!-- If you are new to the Delta-V repository, please read the [Contributing Guidelines](https://github.com/DeltaV-Station/Delta-v/blob/master/CONTRIBUTING.md) -->

## About the PR
* Magnet pull time has been increased from 6 min to 9 min.
* Magnet release warning message time has been increased from 30 seconds to 59 seconds.
* Magnet offer refresh time has been decreased from 3 min to 2 min.
* Probability of wrecks for the salvage magnet has been increased, and probability of asteroids has been decreased.
* Lowered the cost to pull wrecks from 1000 mining points to 500 mining points, to put in line with nerfs to ore output and improved materials from recycling scrap.
* Time for a fulton to teleport has been decreased from 45 seconds to 30 seconds.

## Why / Balance
* Currently there is just not enough time to reasonably explore and salvage wrecks - only 6 minutes are given. 9 minutes represents a 50% increase in time.  
* 2 minutes is a 1 minute decrease in time between offers - making it easier to way for a new offer if you want specifically a large wreck, for example.
* Increased the time before magnet pull deletion to give enough time to fulton out one last crate or locker.
* For the same reason as above, decreased time to fulton to 30 seconds from 45 seconds. Only a 15 second difference.
* Decreased point cost to pull wrecks to 500 points. Combined w recycling (and selling) buffs in my other PR, this just makes the "fun tax" lesser. However, it still has a cost to encourage at least round-start mining. Also keep in mind my other PR nerfs ores somewhat, so a point cost decrease is also part of the reason.
* Made wrecks more probable (4.5f) and dropped asteroid probability (2.0f). Asteroids were already very rare due to how much better lavaland was. Currently, asteroids do not present any option (nor do debris), so there is no reason for there to be more than one option. These values can be changed in the future if the mining rework happens.

If I can figure out how to do UI (and proper C#), the future plan is to increase the time further (to really allow to strip a wreck bare and explore it), but add an option to have an early magnet release. This is a stopgap due to my current skills.
If I can figure out how to do UI (and proper C#), in the eventual future plan is also to add mining points to the recycler as well.

## Technical details
Value modifications. Simply changes the timings in three files, plus float values in one.

The time chosen for the warning message (59 seconds) is due to .Seconds component only taking values up to 59 - thus 60 seconds is not possible. I could have used .TotalSeconds, but then it displays a fraction. Womp.

## Media
<!-- Attach media if the PR makes ingame changes (clothing, items, features, etc).
Small fixes/refactors are exempt. -->

## Requirements
<!-- Confirm the following by placing an X in the brackets [X]: -->
- [X] I have tested all added content and changes.
- [X] I have added media to this PR or it does not require an ingame showcase.
<!-- You should understand that not following the above may get your PR closed at maintainer’s discretion -->

## Breaking changes
<!-- List any breaking changes, including namespaces, public class/method/field changes, prototype renames; and provide instructions for fixing them. -->

**Changelog**
<!-- Add a Changelog entry to make players aware of new features or changes that could affect gameplay.
Make sure to read the guidelines and take this Changelog template out of the comment block in order for it to show up.
Changelogs must have a :cl: symbol, so the bot recognizes the changes and adds them to the game's changelog. -->

:cl:
- tweak: Magnet pull time has been increased from 6 min to 9 min.
- tweak: Magnet release warning message time has been increased from 30 seconds to 59 seconds.
- tweak: Magnet offer refresh time has been decreased from 3 min to 2 min.
- tweak: Probability of wrecks for the salvage magnet has been increased, and probability of asteroids has been decreased.
- tweak: Lowered the cost to pull wrecks from 1000 mining points to 500 mining points, to put in line with nerfs to ore output and improved materials from recycling scrap.
- tweak: Time for a fulton to teleport has been decreased from 45 seconds to 30 seconds.