@Avalon-Proto 2025-06-13
I will mention, wrecks are really not worth keeping a full 9 minutes, most of them you explore and strip of loot in a manner of 2-3 minutes, Leaving a whole 6 minutes of wasted time which wrecks can be pulled in