@HTMLSystem 2025-06-13
> I will mention, wrecks are really not worth keeping a full 9 minutes, most of them you explore and strip of loot in a manner of 2-3 minutes, Leaving a whole 6 minutes of wasted time which wrecks can be pulled in

Occasional salv player here... I've never completed a wreck or a debris haha! I would love the extra time, makes me feel like I actually have a chance!