@randomboredguy 2025-06-13
> steal mining stations thing of having no time limit and only releasing by interacting with the magnet

Honestly, best suggestion.

Maybe have wrecks cost by size. small don't have the same wealth pay out as a mid or large so they are very rarely selected. like a 450 for small, 700 for mid and 1200 for large.