@AlgisAlphonse 2025-06-13
> steal mining stations thing of having no time limit and only releasing by interacting with the magnet

Wait, what thing? Genuinely can you link to it, I do not know any C#.